+++ app/src/main/java/com/elchaninov/gbprofessionaldevelopment/data/room/TranslationDao.kt
@@ -6,6 +6,9 @@ import androidx.room.OnConflictStrategy.IGNORE
 import androidx.room.OnConflictStrategy.REPLACE
 import androidx.room.Query
 import androidx.room.Transaction
+import com.elchaninov.gbprofessionaldevelopment.model.room.EntityMeaning
+import com.elchaninov.gbprofessionaldevelopment.model.room.EntityTranslation
+import com.elchaninov.gbprofessionaldevelopment.model.room.EntityTranslationWhitsMeaning
 
 @Dao
 interface TranslationDao {
@@ -26,22 +29,22 @@ interface TranslationDao {
     }
 
     @Query("SELECT * FROM TranslationTable WHERE text LIKE :word ORDER BY text ASC")
-    suspend fun getTranslationWhitsMeaning(word: String): List<TranslationWhitsMeaning>
+    suspend fun getTranslationWhitsMeaning(word: String): List<EntityTranslationWhitsMeaning>
 
     @Query("SELECT * FROM TranslationTable WHERE favorite=:favorite AND text LIKE :word ORDER BY text ASC")
     suspend fun getFavoriteTranslationWhitsMeaning(
         word: String,
         favorite: Boolean = true
-    ): List<TranslationWhitsMeaning>
+    ): List<EntityTranslationWhitsMeaning>
 
     @Query("SELECT * FROM TranslationTable WHERE text = :word")
-    suspend fun getTranslationWhitsMeaningByText(word: String): TranslationWhitsMeaning
+    suspend fun getTranslationWhitsMeaningByText(word: String): EntityTranslationWhitsMeaning
 
     @Query("UPDATE TranslationTable SET favorite = :favorite WHERE text =:word")
     suspend fun updateTranslationFavorite(word: String, favorite: Boolean)
 
     @Transaction
-    suspend fun toggleTranslationFavorite(word: String): TranslationWhitsMeaning {
+    suspend fun toggleTranslationFavorite(word: String): EntityTranslationWhitsMeaning {
         val currentFavorite = getTranslationWhitsMeaningByText(word)
         updateTranslationFavorite(word, !currentFavorite.favorite)
         return getTranslationWhitsMeaningByText(word)
