+++ app/src/main/java/com/elchaninov/gbprofessionaldevelopment/model/usermodel/DataModel.kt
@@ -1,14 +1,7 @@
 package com.elchaninov.gbprofessionaldevelopment.model.usermodel
 
-import com.google.gson.annotations.SerializedName
-
 data class DataModel(
-    @SerializedName("text")
     val text: String?,
-
-    @SerializedName("meanings")
     val meanings: List<Meanings>?,
-
-    @SerializedName("favorite")
     val favorite: Boolean = false
 )
