+++ app/src/main/java/com/elchaninov/gbprofessionaldevelopment/model/usermodel/Meanings.kt
@@ -1,11 +1,6 @@
 package com.elchaninov.gbprofessionaldevelopment.model.usermodel
 
-import com.google.gson.annotations.SerializedName
-
 data class Meanings(
-    @SerializedName("translation")
     val translation: Translation?,
-
-    @SerializedName("imageUrl")
     val imageUrl: String?
 )
