+++ app/src/main/java/com/elchaninov/gbprofessionaldevelopment/model/usermodel/Translation.kt
@@ -1,8 +1,5 @@
 package com.elchaninov.gbprofessionaldevelopment.model.usermodel
 
-import com.google.gson.annotations.SerializedName
-
 data class Translation(
-    @SerializedName("text")
     val translation: String?
 )
