+++ app/src/main/java/com/elchaninov/gbprofessionaldevelopment/view/main/MainInteractor.kt
@@ -3,7 +3,7 @@ package com.elchaninov.gbprofessionaldevelopment.view.main
 import com.elchaninov.gbprofessionaldevelopment.data.DataSourceLocal
 import com.elchaninov.gbprofessionaldevelopment.data.DataSourceRemote
 import com.elchaninov.gbprofessionaldevelopment.model.usermodel.DataModel
-import com.elchaninov.gbprofessionaldevelopment.utils.toDataModel
+import com.elchaninov.gbprofessionaldevelopment.model.toDataModel
 import com.elchaninov.gbprofessionaldevelopment.viewmodel.Interactor
 
 class MainInteractor(
